- Added TextViewElement extension function - Fixed small glitches
