+++ app/src/main/java/com/kickstarter/ui/fragments/projectpage/ProjectCampaignFragment.kt
@@ -37,7 +37,7 @@ class ProjectCampaignFragment : BaseFragment<ProjectCampaignViewModel.ViewModel>
         this.viewModel.outputs.storyViewElements()
             .subscribeOn(Schedulers.io())
             .distinctUntilChanged()
-            .delay(200, TimeUnit.MILLISECONDS)
+            .delay(150, TimeUnit.MILLISECONDS)
             .compose(bindToLifecycle())
             .compose(Transformers.observeForUI())
             .subscribe {
