+++ app/src/main/res/drawable/rectangle_grey.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <size android:height="569dp" android:width="1012dp" />
+    <size android:height="539dp" android:width="1012dp" />
     <gradient
         android:angle="90"
         android:endColor="#f1f3f4"
